// types/opts.d.ts
type AREA_COLOR_VALUE = "Seoul" | "Gyeonggi" | "Sejong" | "Daejeon" | "Daegu" | "Busan" | "Gwangju" | "Jeju" | "Gangwon" | "Gyungnam" | "Gyungbok" | "Ulsan" | "Incheon" | "Jeonnam" | "Jeonbok" | "Chungname" | "Chungbuk";

export interface AREA_COLOR {
    [key : string] : AREA_COLOR_VALUE
}

type TYPES_COLOR_VALUE = "theatre" | "exhibition" | "etc" | "dance" | "music";

export interface TYPES_COLOR {
    [key : string] : TYPES_COLOR_VALUE
}